Split functionality out into appropriate lib/bin spaces
