+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "lint": "eslint .",
-    "start": "node index.js",
+    "start": "orglinter",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -23,5 +23,8 @@
   "devDependencies": {
     "eslint": "^7.15.0",
     "eslint-config-godaddy": "^4.0.1"
+  },
+  "bin": {
+    "orglinter": "./src/bin/cli.js"
   }
 }
